Initial CLI refactoring and test infra
This is the first piece of work on the testing
system that the CLI will soon have. The current
method relies on replacing the dependencies of
the CLI with mock versions of the actual thing.

Another big part of the work in this
patch is around better error propagation
and handling in the cli-sdk. Furthermore
the docker setup was improved to allow
us to run as a non root user.